mlx4: Add mlx4 direct verbs
Direct verbs feature provides fast data path execution
by implementing verbs directly over hardware specification.

It consists from the two parts:
 * Exported functions to translate internal mlx4 structures to be
   accessible by user applications.
 * Inline functions, enums and defines to simplify programmer's life.

This patch adds one header file (mlx4dv.h) with general, CQE and WQE data.
